tests: move assert_args_and_raise and assert_args_and_return to helpers.py for broader use
